+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.10.1
+### Bugfix
+* fix glow effect of prompt when it don't have formatting
+* fix exec commands from echo
+* fix echo formatting with newline and bracket at the end
+* fix single wide character exception
+* fix substring after string when string have bracket at the end [#550](https://github.com/jcubic/jquery.terminal/issues/550)
+* fix images & style from formatters inside cmd [#519](https://github.com/jcubic/jquery.terminal/issues/519)
+
 ## 2.10.0
 ### Feature
 * improve performance by adding function `$.terminal.partition`